refactor(outputview, cars): 시스템 출력과 비즈니스 로직 분리
